[batch] Move phase module from batch to the services folder and rename to batch
