+++ src/batch/harvest.ts
@@ -3,6 +3,7 @@ import type { AutocompleteData, NS } from "netscript";
 import { HostAllocation, MemoryClient, registerAllocationOwnership } from "services/client/memory";
 import { TAG_ARG } from "services/client/memory_tag";
 import { PortClient } from "services/client/port";
+
 import { TaskSelectorClient, Lifecycle } from "batch/client/task_selector";
 
 import { CONFIG } from "batch/config";
@@ -12,7 +13,8 @@ import {
     fullBatchTime,
     growthAnalyze
 } from "batch/expected_value";
-import { BatchLogistics, BatchPhase, calculatePhaseStartTimes, spawnBatch } from "batch/phase";
+
+import { BatchLogistics, BatchPhase, calculatePhaseStartTimes, spawnBatch } from "services/batch";
 
 
 export function autocomplete(data: AutocompleteData, _args: string[]): string[] {
