+++ src/batch/sow.ts
@@ -7,11 +7,12 @@ import {
     MemoryClient,
     TransferableAllocation,
 } from "services/client/memory";
+import { TAG_ARG } from "services/client/memory_tag";
 
 import { CONFIG } from "batch/config";
-import { BatchLogistics, BatchPhase, calculatePhaseStartTimes } from "batch/phase";
 import { awaitRound, calculateRoundInfo, RoundInfo } from "batch/progress";
-import { TAG_ARG } from "services/client/memory_tag";
+
+import { BatchLogistics, BatchPhase, calculatePhaseStartTimes } from "services/batch";
 
 const GROW_SCRIPT = "/batch/g.js";
 const WEAKEN_SCRIPT = "/batch/w.js";
